GUI层框架完成 需理清逻辑 从底层开始编码

+++ hqtop/systemdataprovider.h
@@ -19,7 +19,7 @@ public:
     // 获取系统总体资源
     virtual SystemResource getSystemResource() = 0;
     // 杀死进程
-    virtual bool killProcess(qint64 pid) = 0;
+    virtual bool killProcess(int pid) = 0;
 
 
     // 析构函数
